address new review comments
