+++ lib/oli_web/router.ex
@@ -1612,7 +1612,7 @@ defmodule OliWeb.Router do
       end
 
       # External tools
-      live("/external_tools", Admin.ExternalToolsView)
+      live("/external_tools", Admin.ExternalTools.ExternalToolsView)
     end
 
     # System admin
