+++ test/oli_web/live/admin/external_tools/external_tools_view_test.exs
@@ -1,4 +1,4 @@
-defmodule OliWeb.Admin.ExternalToolsViewTest do
+defmodule OliWeb.Admin.ExternalTools.ExternalToolsViewTest do
   use OliWeb.ConnCase, async: true
   import Phoenix.LiveViewTest
   import Oli.Factory
